+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased changes (in master but not released in a version)
 
+- [FileOrganizer] Passes isShownOnLoad to thumbnail render () @liamross
+- [useManagedFiles] Hook for managing files for FileOrganizer () @liamross
 - [FileOrganizer] Virtualized if many files (#4) @liamross
 - [File] Implemented MemoizedPromise (#3) @liamross
 - [Thumbnail] Fixed space key not working for renaming (#2) @liamross